pyproject.toml
[build-system]
requires = [
    "hatchling>=1.27.0",
]
build-backend = "hatchling.build"

[project]
name = "checkdmarc"
dynamic = [
    "version",
]
python_requires = ">=3.9"
description = "A Python module and command line parser for SPF and DMARC records"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Sean Whalen", email = "whalenster@gmail.com" },
]
keywords = [
    "DMARC",
    "DNS",
    "SPF",
    "BIMI",
    "MTA-STS"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Security",
    "Topic :: Communications :: Email",
]
dependencies = [
    "dnspython>=2.0.0",
    "cryptography>=45.0,<47.0",
    "pyopenssl>=24.2.1",
    "pem>=23.1.0",
    "expiringdict>=1.1.4",
    "publicsuffixlist>=0.10.0",
    "pyleri>=1.3.2",
    "requests>=2.25.0",
    "timeout-decorator>=0.4.1",
    "xmltodict>=0.14.2",
    "importlib-resources>=6.0",
]

[project.optional-dependencies]
build = [
    "hatch>=1.14.0",
    "myst-parser[linkify]",
    "nose",
    "pytest",
    "pytest-cov",
    "ruff",
    "sphinx",
    "sphinx_rtd_theme",
]


[project.scripts]
checkdmarc = "checkdmarc._cli:_main"

[project.urls]
Homepage = "https://github.com/domainaware/checkdmarc"
Documentation = "https://domainaware.github.io/checkdmarc/"
Issues = "https://github.com/domainaware/checkdmarc/issues"
Changelog = "https://github.com/domainaware/checkdmarc/blob/master/CHANGELOG.md"

[tool.hatch.version]
path = "checkdmarc/_constants.py"

[tool.hatch.build.targets.sdist]
include = [
    "/checkdmarc",
]

[tool.hatch.build.targets.wheel]
packages = ["checkdmarc"]
